+++ R/blav_test.R
@@ -11,6 +11,7 @@ blav_model_test <- function(lavmodel       = NULL,
                             samplls        = NULL,
                             jagextra       = NULL,
                             stansumm       = NULL,
+                            domll          = NULL,
                             control        = list()) {
 
 
@@ -21,14 +22,14 @@ blav_model_test <- function(lavmodel       = NULL,
     if("syntax" %in% names(jagextra)){
         warning("blavaan WARNING: Marginal log-likelihood cannot be approximated when there is additional JAGS syntax.", call. = FALSE)
         mll <- NA
-    } else if(lavoptions$categorical){
-        mll <- NA # not tested, priors may cause problems
-    } else {
+    } else if(domll) {
         mll <- try(margloglik(lavpartable, lavmodel, lavoptions, 
                               lavsamplestats, lavdata, lavcache,
                               lavjags, VCOV, x, stansumm),
                    silent=TRUE)
         if(inherits(mll, "try-error")) mll <- NA
+    } else {
+        mll <- NA # not tested, priors may cause problems
     }
 
     if(lavoptions$target == "stan") {