+++ R/blav_utils.R
@@ -377,3 +377,91 @@ Mp2dataidx <- function(Mp, case.idx, exclude.empty = TRUE){
 
   out
 }
+
+## check for restricted model covariance matrices, which causes problems for priors
+checkcovs <- function(lavobject){
+  free <- lavInspect(lavobject, 'free')
+
+  ## ensure each list entry is one group
+  if (inherits(free[[1]], "matrix")) free <- list(free)
+
+  if (nrow(free[[1]]$psi) > 0) {
+    psis <- lapply(free, function(x) x$psi)
+    psinums <- sapply(psis, function(x) x[lower.tri(x)])
+    diagpsi <- all(psinums == 0L, na.rm = TRUE)
+    fullpsi <- all(psinums > 0L, na.rm = TRUE) & (anyDuplicated(psinums, MARGIN = 0) == 0L)
+  } else {
+    diagpsi <- FALSE
+    fullpsi <- TRUE
+  }
+
+  if (nrow(free[[1]]$theta) > 0) {
+    thets <- lapply(free, function(x) x$theta)
+    thetnums <- sapply(thets, function(x) x[lower.tri(x)])
+    diagthet <- all(thetnums == 0L, na.rm = TRUE)
+    ## surprising if this happens:
+    fullthet <- all(thetnums > 0L, na.rm = TRUE) & (anyDuplicated(thetnums, MARGIN = 0) == 0L)
+  } else {
+    diagthet <- FALSE
+    fullthet <- TRUE
+  }
+
+  list(diagpsi = diagpsi, fullpsi = fullpsi, diagthet = diagthet, fullthet = fullthet)
+}
+
+## check whether model cov matrix is block diagonal
+## uses matrices from lavInspect(, "free")
+blkdiag <- function(mat) {
+  isblk <- TRUE
+  cnum <- 1L
+  nblks <- 0
+  matrows <- NROW(mat)
+  blkse <- matrix(0, matrows, 3) # start/end of each block, and first free param in that block
+  while (isblk && cnum <= matrows) {
+    ## ending row of this potential block
+    currend <- which(mat[, cnum] > 0)
+    if (length(currend) > 0) {
+      currend <- max(currend)
+    } else {
+      ## a fixed diagonal entry, which is its own block
+      currend <- cnum
+    }
+
+    if (currend > cnum) {
+      ## check that columns cnum+1 to max() also equal max()
+      othend <- sapply((cnum + 1):currend, function(i) {
+        nzents <- which(mat[,i] > 0)
+        if (length(nzents) > 0) {
+          out <- max(nzents)
+        } else {
+          out <- i
+        }
+        out})
+        
+      if (all(othend == currend)) {
+        ## is this entire submatrix free?
+        submat <- mat[cnum:currend, cnum:currend]
+        if (all(submat[lower.tri(submat)] > 0)) {
+          nblks <- nblks + 1
+          blkse[nblks,] <- c(cnum, currend, min(submat[lower.tri(submat)]))
+          cnum <- currend + 1
+        } else {
+          isblk <- FALSE
+        }
+      } else {
+        isblk <- FALSE
+      }
+    } else if (currend == cnum) {
+      ## 1x1 block
+      nblks <- nblks + 1
+      blkse[nblks,] <- c(cnum, cnum, 0)
+      cnum <- cnum + 1
+    } else {
+      isblk <- FALSE
+    }
+  }
+
+  if (isblk) blkse <- blkse[1:nblks, , drop = FALSE]
+  
+  list(isblk = isblk, nblks = nblks, blkse = blkse)
+}