+++ NEWS.md
@@ -1,4 +1,22 @@
-# Version 0.4-2
+# Version 0.4-5
+## New features
+* For target = "stan", meanstructure=FALSE is allowed, along with use of sample.cov and sample.nobs instead of raw data
+
+* Users are warned about priors on covariance matrices that are neither diagonal nor unrestricted
+
+* For models where observed variable intercepts appear in the latent intercept vector (alpha), default priors come from the observed intercept vector nu (as the user would expect)
+
+* inits = "simple" is now default (instead of "prior"), to address some convergence problems
+
+* For stan targets, ":=" can now be used as an identity function
+
+* For target = "stan", fix the missing data issue from 0.4-3 (complete data in one group but not the other)
+
+* Column names are added to blavPredict(, type="lv")
+
+
+
+# Version 0.4-3
 ## New features
 * For target = "stan", most models should run faster than they did in earlier versions (use of sufficient statistics)
 
@@ -15,6 +33,8 @@
 * loo() moment matching available by passing mcmcextra = list(data = list(moment_match_k_threshold))
 
 ## Bugs/glitches discovered after the release:
+* target = "stan" fails when there are complete data in one group and missing data in another group
+
 * blavPredict(, type="ymis") still not available for models with ordinal variables